Renamed extract-jwt to jwt-utils
