+++ lib/decorators/all.js
@@ -1,6 +1,6 @@
 'use strict';
 
-const extractJWT = require('../extract-jwt').extractJWT;
+const extractJWT = require('../jwt-utils').extractJWT;
 
 function decorator() {
     this.request(extractJWT);
