+++ lib/routes/account-linking.js
@@ -2,7 +2,7 @@
 
 const App       = require('../models').App;
 const AppUser   = require('../models').AppUser;
-const decodeJwt = require('../extract-jwt').decodeJWT;
+const decodeJwt = require('../jwt-utils').decodeJWT;
 
 function validateUserId(userProfile, id) {
     return new Promise((resolve, reject) => {
